Add batch size parameter to generate_tfrecord function and scripts
The main change in this commit is the addition of a batch size parameter to the generate_tfrecord function in the brain generator. It also includes changes across multiple scripts and config files to accommodate this new functionality. The introduction of the batch size parameter will allow users to specify the number of training pairs to put into a single tfrecord file, improving fine-grained control over data generation. We also added a new script, tfrecord_to_nifty, and updated some relevant metadata information.